Replace Taskr with the Rust build runner
