+++ .github/workflows/build_and_deploy.yml
@@ -180,6 +180,9 @@ jobs:
           node scripts/set-preview-version.js "${{ github.sha }}"
           pnpm install --no-frozen-lockfile
 
+      - name: Install Rust for the build runner
+        uses: ./.github/actions/setup-rust
+
       - run: pnpm run build
 
       - uses: actions/cache@27d5ce7f107fe9357f9df03efb73ab90386fccae # v5.0.5
