+++ package.json
@@ -146,7 +146,6 @@
     "@swc/core": "1.11.24",
     "@swc/helpers": "0.5.23",
     "@swc/types": "0.1.7",
-    "@taskr/esnext": "1.1.0",
     "@testing-library/jest-dom": "6.1.2",
     "@testing-library/react": "^15.0.5",
     "@types/busboy": "1.5.3",
@@ -286,7 +285,6 @@
     "styled-jsx-plugin-postcss": "3.0.2",
     "swr": "^2.2.4",
     "tailwindcss": "3.2.7",
-    "taskr": "1.1.0",
     "tree-kill": "1.2.2",
     "tsec": "0.2.1",
     "tsx": "4.19.2",
@@ -336,7 +334,6 @@
     "patchedDependencies": {
       "webpack-sources@3.2.3": "patches/webpack-sources@3.2.3.patch",
       "stacktrace-parser@0.1.10": "patches/stacktrace-parser@0.1.10.patch",
-      "taskr@1.1.0": "patches/taskr@1.1.0.patch",
       "minizlib@3.1.0": "patches/minizlib@3.1.0.patch",
       "@rspack/core@1.6.7": "patches/@rspack__core@1.6.7.patch",
       "@modelcontextprotocol/sdk": "patches/@modelcontextprotocol__sdk.patch",
