+++ packages/next-taskr/package.json
@@ -6,6 +6,6 @@
     "next-taskr": "bin/next-taskr.js"
   },
   "scripts": {
-    "build:runner": "node build.js"
+    "build": "node build.js"
   }
 }
