+++ packages/next/build-tools/ncc.js
@@ -3,6 +3,7 @@ const ncc = require('@vercel/ncc')
 const { existsSync, readFileSync } = require('fs')
 const { basename, dirname, extname, join, resolve } = require('path')
 const { Module } = require('module')
+const root = resolve(__dirname, '..')
 
 // files might be lower case and not able to be found on case-sensitive
 // file systems (ubuntu)
@@ -14,54 +15,51 @@ const potentialLicenseFiles = [
   'license.md',
 ]
 
-// See taskfile.js bundleContext definition for explanation
-const m = new Module(resolve(__dirname, 'bundles', '_'))
+// See recipes.js bundleContext definition for explanation
+const m = new Module(resolve(root, 'bundles', '_'))
 m.filename = m.id
 m.paths = Module._nodeModulePaths(m.id)
 const bundleRequire = m.require
 bundleRequire.resolve = (request, options) =>
   Module._resolveFilename(request, m, false, options)
 
-module.exports = function (task) {
-  // eslint-disable-next-line require-yield
-  task.plugin('ncc', {}, function* (file, options) {
-    if (options.externals && options.packageName) {
-      options.externals = { ...options.externals }
-      delete options.externals[options.packageName]
-    }
-    let precompiled = options.precompiled !== false
-    delete options.precompiled
+module.exports = async function (file, options) {
+  if (options.externals && options.packageName) {
+    options.externals = { ...options.externals }
+    delete options.externals[options.packageName]
+  }
+  let precompiled = options.precompiled !== false
+  delete options.precompiled
 
-    return ncc(join(__dirname, file.dir, file.base), {
-      filename: file.base,
-      minify: options.minify === false ? false : true,
-      assetBuilds: true,
-      cache: false,
-      ...options,
-    }).then(({ code, assets }) => {
-      Object.keys(assets).forEach((key) => {
-        let data = assets[key].source
+  return ncc(join(root, file.dir, file.base), {
+    filename: file.base,
+    minify: options.minify === false ? false : true,
+    assetBuilds: true,
+    cache: false,
+    ...options,
+  }).then(({ code, assets }) => {
+    Object.keys(assets).forEach((key) => {
+      let data = assets[key].source
 
-        this._.files.push({
-          data,
-          base: basename(key),
-          dir: join(file.dir, dirname(key)),
-        })
+      this._.files.push({
+        data,
+        base: basename(key),
+        dir: join(file.dir, dirname(key)),
       })
-
-      if (options && options.packageName) {
-        writePackageManifest.call(
-          this,
-          options.packageName,
-          file.base,
-          options.bundleName,
-          precompiled,
-          options.packageJsonName
-        )
-      }
-
-      file.data = Buffer.from(code, 'utf8')
     })
+
+    if (options && options.packageName) {
+      writePackageManifest.call(
+        this,
+        options.packageName,
+        file.base,
+        options.bundleName,
+        precompiled,
+        options.packageJsonName
+      )
+    }
+
+    file.data = Buffer.from(code, 'utf8')
   })
 }
 
@@ -89,7 +87,7 @@ function writePackageManifest(
   let { name, author, license } = require(packagePath)
 
   const compiledPackagePath = join(
-    __dirname,
+    root,
     `${!precompiled ? 'dist/' : ''}src/compiled/${bundleName || packageName}`
   )
