+++ packages/next/package.json
@@ -81,26 +81,15 @@
     "next": "./dist/bin/next"
   },
   "scripts": {
-    "dev": "cross-env NEXT_SERVER_NO_MANGLE=1 taskr",
-    "build": "taskr release",
-    "build:rust": "node ../next-taskr/bin/next-taskr.js release",
-    "dev:rust": "cross-env NEXT_SERVER_NO_MANGLE=1 node ../next-taskr/bin/next-taskr.js",
-    "ncc-compiled:rust": "node ../next-taskr/bin/next-taskr.js ncc",
+    "dev": "cross-env NEXT_SERVER_NO_MANGLE=1 next-taskr",
+    "build": "next-taskr release",
     "types": "tsc --project tsconfig.build.json --declaration --emitDeclarationOnly --stripInternal --declarationDir dist",
     "typescript": "tsec --noEmit",
-    "ncc-compiled": "taskr ncc",
+    "ncc-compiled": "next-taskr ncc",
     "storybook": "BROWSER=none storybook dev -p 6006",
     "build-storybook": "storybook build",
     "test-storybook": "test-storybook",
     "pack-for-isolated-tests": "pnpm pack --out ./packed.tgz"
-  },
-  "taskr": {
-    "requires": [
-      "./taskfile-webpack.js",
-      "./taskfile-ncc.js",
-      "./taskfile-swc.js",
-      "./taskfile-watch.js"
-    ]
   },
   "dependencies": {
     "@next/env": "16.4.0-canary.20",
@@ -172,6 +161,7 @@
     "@next/polyfill-nomodule": "16.4.0-canary.20",
     "@next/react-refresh-utils": "16.4.0-canary.20",
     "@next/swc": "16.4.0-canary.20",
+    "@next/taskr": "workspace:*",
     "@opentelemetry/api": "1.6.0",
     "@playwright/test": "1.61.0",
     "@rspack/core": "1.6.7",
@@ -186,8 +176,6 @@
     "@storybook/test-runner": "0.21.0",
     "@swc/core": "1.11.24",
     "@swc/types": "0.1.7",
-    "@taskr/clear": "1.1.0",
-    "@taskr/esnext": "1.1.0",
     "@types/babel__code-frame": "7.0.6",
     "@types/babel__core": "7.20.5",
     "@types/babel__generator": "7.27.0",
@@ -342,7 +330,6 @@
     "style-loader": "4.0.0",
     "superstruct": "1.0.3",
     "tar": "7.5.11",
-    "taskr": "1.1.0",
     "terser": "5.27.0",
     "terser-webpack-plugin": "5.3.9",
     "text-table": "0.2.0",
