+++ scripts/publish-release.js
@@ -220,7 +220,7 @@ const publishRetryDelaySeconds = 15
   // before publishing to reduce the total metadata size. The max is
   // 100 MB for all versions of a pkg and we are currently at 30MB. Run:
   // curl -w '%{size_download}' -so /dev/null https://registry.npmjs.org/next | awk '{print $1/1000000 " MB"}'
-  for (const field of ['devDependencies', 'taskr', 'scripts']) {
+  for (const field of ['devDependencies', 'scripts']) {
     delete nextPkg[field]
   }
   await fs.writeFile(
