+++ test/unit/next-taskr/fixtures/no-bundlers.cjs
@@ -1,7 +1,12 @@
 const Module = require('node:module')
 const load = Module._load
 Module._load = function (request, ...args) {
-  if (request === '@vercel/ncc' || request === '@rspack/core') {
+  if (
+    request === '@vercel/ncc' ||
+    request === '@rspack/core' ||
+    request === 'taskr' ||
+    request.startsWith('@taskr/')
+  ) {
     throw new Error(`A cached SWC recipe must not load ${request}`)
   }
   return load.call(this, request, ...args)
